Use hvl2wav replay as base for HivelyTracker
hopefully fixes Linux/ARM64 garbled audio...
