+++ src/songend/Makefile.am
@@ -34,10 +34,8 @@ uade_songend_SOURCES = \
 lib3rdparty_a_SOURCES = \
     ../3rdparty/md5.h \
     ../3rdparty/md5.cc \
-    ../3rdparty/hvl/hvl_replay.h \
-    ../3rdparty/hvl/hvl_replay.c \
-    ../3rdparty/hvl/hvl_tables.h \
-    ../3rdparty/hvl/hvl_tables.c \
+    ../3rdparty/hvl/replay.h \
+    ../3rdparty/hvl/replay.c \
     ../3rdparty/dbm/dsp.h \
     ../3rdparty/dbm/dsp_echo.c \
     ../3rdparty/dbm/dsp_fetchinstr.c \
